refactor(database/reminders): add FK relationship to user
this adds a one-to-many relationship for reminders & users.

this change also exposes a lot of issues with the current system of entities, and it should be changed (as described earlier in the PR) as it does introduce a dangerous cast and basically needs to bypass ReminderData & UserData as relationships require a ORMLITE entity to work.

personally, I think we should just add functionality to ORMLite or use it as a base for exotic database types rather than the current system.

however, this can be discussed more in depth in an issue